+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [ADDED] `Sequelize.prototype.loadModels` method
+
 ## 0.1.0
 
 * [ADDED] initial release - extracted from bi-service@0.15.3 (no additional changes to the logic)